Add files via upload
To understand global economic inequalities, a comparison between developed and undeveloped nation states is necessary. This research seeks to investigate underlying reasons for such differences by comparing 20 developed and 20 developing countries. This comparison is rooted on various economic and non-economic measures such as GDP, life expectancy, rate of inflation, and literacy levels. In this way, we use statistical approaches to find some relations in below objectives:
1. Compare GDP, GDP per capita and GDP growth rates between developed and underdeveloped countries. 
2. Find most important factors which impacting on inflation rates and the wight of them. As well as this, showing does    inflation can impact on other economic sections or not. If yes, how and why?
3. Correlation between GDP per capita and Life Expectancy in developed countries.
4. Find the relation between access to electricity and life expectancy. 
5. Investigate on relations between deposit/lending rates and inflation.
6. Finding relation and the strength of relation between mortality infant or access to electricity and life expectancy. 
7. Predict GDP and GDP per capita trend in developed and undeveloped countries. 
The overall goal is to conduct comparative statistical analysis to identify relationships and differences between the two groups in order to better understand the relationships of factors and also the impact of them on other subjects like economic growth and quality of life